Bug Fix: ordered calls result in unordered retrieval
In the event that a previous call to retrieve a message or a response
timesout, the next call must retrieve the latest message, ignoring the
previously timed out retrievals. If there are three calls to send a
message for processing, and the second call to retrieve timesout, the
third call to retrieve should discard the second call, and retrieve the
third message instead.