Security Fix: let KeyProvider handle the quick loading of keys
Best practice demands that keys be kept in memory for the shortest
possible time. Ideal scenarios will use hardware keys which do not
provide key material but a link to the hardware key. For software
protected keys, the keys should live in memory for the shortest amount
of time possible.

Using KeyProvider to retrieve the per encryption operation key allows
the keyprovider to handle details such as wiping key material from
memory (via System.array utilities, for ex), whiles still maintaining
the necessary optimization such as reducing retrieval times to reduce
the overall cryptographic operations.

The standard key provider simply retrieves the key from a keystore file.
The retrieved key is garbage collected (potentially after a long time)
which makes it not ideal for high security environments, and quite slow.
However unlike the previous versions the key is not long lived in memory

Optionally the destroy() key method for Destroyable keys is called. This
can provide some level of guarantee that the key material has been
destroyed, though with Java, that is not guaranteed.