Fix: Cache De-Serialization fails due to white list errors
Added all relevant classes in serialization down the object attribute
tree for ProcessingLog and ResultExceptionLog. This is due to
Infinispan's JavaSerialization whitelist designed to prevent an attacker
from causing the deserialization of a malicious payload.

For Syshub, all de-serialized data comes from a controlled, trusted
source, the Logging Database. As long as this database is secured, the
ability of an attacker to poison the system with a malicious payload is
eliminated. The serialized data is also not derived from user input,
thus is safe from tampering.